Allow all clippy lints for generated code.
The lints are generally appropriate for hand-written code, but
oftentimes the best solution for generated code violates such
guidelines. Typically, therefore, generated code is not subject to
such linting.

The most precise way to disable linting for this generated code in
a bazel environment is to add `#![allow(clippy::all)]` to the top
level of the generated crate.